Move common code from analyzer to 'common' folder.
