+++ src/main/java/com/autotune/analyzer/k8sObjects/ValidateAutotuneObject.java
@@ -15,9 +15,9 @@
  *******************************************************************************/
 package com.autotune.analyzer.k8sObjects;
 
-import com.autotune.analyzer.utils.AutotuneSupportedTypes;
-import com.autotune.analyzer.utils.AnalyzerConstants;
-import com.autotune.analyzer.utils.AnalyzerErrorConstants;
+import com.autotune.utils.AutotuneSupportedTypes;
+import com.autotune.utils.AnalyzerConstants;
+import com.autotune.utils.AnalyzerErrorConstants;
 
 import java.util.HashMap;
