+++ src/main/java/com/autotune/analyzer/layer/ContainerLayer.java
@@ -1,8 +1,8 @@
 package com.autotune.analyzer.layer;
 
 import com.autotune.analyzer.application.Tunable;
-import com.autotune.analyzer.experiments.PodContainer;
-import com.autotune.analyzer.experiments.Resources;
+import com.autotune.common.data.experiments.PodContainer;
+import com.autotune.common.data.experiments.Resources;
 import org.json.JSONObject;
 
 public class ContainerLayer extends GenericLayer implements Layer {
