+++ src/main/java/com/autotune/analyzer/utils/ExperimentHelpers.java
@@ -2,7 +2,7 @@ package com.autotune.analyzer.utils;
 
 import com.autotune.analyzer.AutotuneExperiment;
 import com.autotune.analyzer.application.ApplicationServiceStack;
-import com.autotune.analyzer.experiments.ExperimentTrial;
+import com.autotune.common.data.experiments.ExperimentTrial;
 import com.autotune.analyzer.k8sObjects.AutotuneConfig;
 import org.json.JSONArray;
 import org.json.JSONObject;
