+++ src/main/java/com/autotune/analyzer/utils/ServiceHelpers.java
@@ -23,6 +23,7 @@ import com.autotune.analyzer.deployment.AutotuneDeploymentInfo;
 import com.autotune.analyzer.k8sObjects.AutotuneConfig;
 import com.autotune.analyzer.k8sObjects.AutotuneObject;
 import com.autotune.analyzer.k8sObjects.Metric;
+import com.autotune.utils.AnalyzerConstants;
 import org.json.JSONArray;
 import org.json.JSONObject;
