+++ src/main/java/com/autotune/common/data/datasource/DataSource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.autotune.analyzer.datasource;
+package com.autotune.common.data.datasource;
 
 import java.net.MalformedURLException;
 import java.util.List;
