+++ src/main/java/com/autotune/queue/AutotuneDTO.java
@@ -20,7 +20,7 @@ import java.io.Serializable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.autotune.analyzer.datasource.DataSourceFactory;
+import com.autotune.common.data.datasource.DataSourceFactory;
 import com.autotune.utils.AutotuneUtils;
 
 /**
@@ -38,16 +38,16 @@ public class AutotuneDTO implements Serializable {
 	private StringBuffer infoMessage;
 	private StringBuffer errorMessage;
 	private static final Logger LOGGER = LoggerFactory.getLogger(DataSourceFactory.class);
-	
+
 	public AutotuneDTO() {}
-	
+
 	/**
-	 * 
+	 *
 	 * @param experimentId : Id generated by Dependency analyzer
 	 * @param componentName: Name of the component processing this DTO
 	 * @param dataURL: fetch the data(JSON)from recommendation manager using this url
 	 * @param operation : set operation name using Operation enum from AutotuneUtil.java
-	 *  
+	 *
 	 */
 	public AutotuneDTO(int experimentId, String componentName, String dataURL, String operation) {
 		this.id = experimentId;
@@ -57,19 +57,19 @@ public class AutotuneDTO implements Serializable {
 		this.errorMessage = new StringBuffer();
 		this.operation = operation;
 	}
-	
+
 	public int getId() {
 		return id;
 	}
-	
+
 	public void setId(int id) {
 		this.id = id;
 	}
-	
+
 	public String getName() {
 		return name;
 	}
-	
+
 	public void setName(String name) {
 		this.name = name;
 	}
@@ -85,7 +85,7 @@ public class AutotuneDTO implements Serializable {
 		}
 		this.url = url;
 	}
-	
+
 	public StringBuffer getInfoMessage() {
 		return infoMessage;
 	}
@@ -101,7 +101,7 @@ public class AutotuneDTO implements Serializable {
 	public void setErrorMessage(StringBuffer errorMessage) {
 		this.errorMessage = errorMessage;
 	}
-	
+
 
 	public String getOperation() {
 		return operation;
@@ -110,7 +110,7 @@ public class AutotuneDTO implements Serializable {
 	public void setOperation(String operation) {
 		this.operation = operation;
 	}
-	
+
 
 	@Override
 	public String toString() {
