+++ src/main/java/com/autotune/utils/HttpUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.autotune.analyzer.utils;
+package com.autotune.utils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
